Split aircraft and background map updates. Add APT database.
There was a *big* performance hit when we added the airport database.
Two things need to happen: (1) we should only update the background
objects when the map has changed, and (2) we should only display data
that is actually visible.  This commit addresses the first fix.  We now
only update the background on a map update.